Refactor: 패키지 리팩토링 진행 및 클래스명 변경
CarRacingGame -> CarRacingController 이름 변경
